Update ignore.lua
Additional Game Event Triggers added for Example Whistle this will help up to remove the need to loop for calling Horse. Also Player Prompt to be able to utilize things like the Emote Menu Prompt to Actually open the Emote menu. the data.complete is now sent if struct size is greater than 2.